pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.23,<0.4.0)",
    "supabase (>=2.15.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "langchain-community (>=0.3.25,<0.4.0)",
    "langchain-core (>=0.3.65,<0.4.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "transformers (>=4.53.0,<5.0.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"

